==> pandas/02.DataFrame.ipynb <==
2.Data Frame
2차원 데이터 (Series들의 모음) 

Data 준비
사전(dict) 자료 구조를 통해 생성 

data = {
    '이름' : ['채치수', '정대만', '홍길동', '김갑수'],
    '학교' : ['북산고', '북산고', '서울고', '서울고'],
    '키' : ['190','180','170','180'],
    '특기' : ['자바', '파이썬', '', '자바']
}
data

data['이름']

import pandas as pd
df = pd.DataFrame(data)

df

df['이름']

df[['이름','키']]

DataFrame 객체 생성 (index 지정)

df = pd.DataFrame(data, index=['1번', '2번','3번','4번'])
df

DataFrame 객체 생성 (Column 지정)
data 중에서 원하는 column만 선택하거나 순서 변경 가능 

df = pd.DataFrame(data, columns=['이름', '학교','키'])
df

df.index

Index 이름 설정

df.index.name = '지원번호'
df
