==> pandas/04.FileSaveOpen.ipynb <==
4.파일 저장 및 열기
#Data Frame 객체를 excel, csv, txt 등 형태의 파일로 저장 

import pandas as pd 
data = {
    '이름' : ['채치수', '정대만', '홍길동', '김갑수'],
    '학교' : ['북산고', '북산고', '서울고', '서울고'],
    '키' : ['190','180','70','80'],
    '국어' : ['90','80','70','80'],
    '영어' : ['90','80','70','80'],
    '수학' : ['90','80','70','80'],
    '특기' : ['자바', '파이썬', '', '자바']
}
df = pd.DataFrame(data, index=['1번', '2번','3번','4번'])
df.index.name = '지원번호'
df

# 저장하기 
csv 파일로 저장

df.to_csv('score1.csv', encoding='utf-8-sig')

df.to_csv('score.csv', encoding='utf-8-sig', index=False)  #인덱스 제외

df.to_csv('score.csv', sep='\t')   #tab으로 구분된 텍스트 파일 

df = pd.read_csv('score.csv', skiprows=2, nrows=4)  #처음 2행은 무시, 이후 4행 가져옴 
df

#df.to_excel('score.xlsx')      # excel 파일로 저장 

df = pd.read_csv('score.csv')
df

df = pd.read_csv('score.csv', skiprows=1, nrows=3)  #처음 1row제외, 이후 3row 가져옴 
df
