==> pandas/05.DataCheck.ipynb <==
# 데이터 확인, 데이터프레임 확인

import pandas as pd 
df = pd.read_csv('score.csv')
df

df.describe()

df.info()

df.head()   #처음 5개의 데이터를 가져옴

df.tail(3)   #마지막 3개의 데이터를 가져옴

df.shape   #8행 10열

df.values

df['키'].describe()

df.to_csv('score1.csv', encoding='utf-8-sig')

df.to_csv('score2.csv', encoding='utf-8', index=False)

df.to_csv('score.txt', sep='\t')  #tab seperator

df.to_csv('score1.xls', encoding='utf-8-sig')

df['키'].describe

df['키'].mean()

df['키'].min()
