==> pandas/06.DataSelect.ipynb <==
# 필요한 라이브러리 임포트
pandas 라이브러리를 임포트합니다.

# pandas 라이브러리를 임포트합니다.
import pandas as pd

# score.txt 파일 읽기
pandas의 read_csv 함수를 사용하여 score.txt 파일을 읽습니다.

# score.txt 파일 읽기
df = pd.read_csv('score.txt', delimiter='\t')  # score.txt 파일을 읽습니다.
height_data = df['키']  # '키' 칼럼의 데이터를 읽어옵니다.
height_data.head()  # 데이터의 일부를 출력합니다.

# '키' 칼럼 데이터 추출
데이터프레임에서 '키' 칼럼의 데이터를 추출합니다.

# '키' 칼럼 데이터 추출
height_data = df['키']  # '키' 칼럼의 데이터를 읽어옵니다.
height_data.head()  # 데이터의 일부를 출력합니다.

df.head()

df['이름']

df[['이름','키']]

# 칼럼 선택 (label)

df.columns

df.columns[2]

df[df.columns[0]]    # [-1] 이면 맨끝에 있는 칼럼명 

df['영어'][0:5]   #0-4까지의 영어점수를 가져옴 

df[['이름','키']][:3]   #처음 3명의 키, 이름 가져옴
