==> pandas/07.DataLoc.ipynb <==
# 07.데이터 선택 (loc)
이름을 이용해서 원하는 행에서 원하는 컬럼 선택

import pandas as pd 
df = pd.read_csv('score.csv', index_col='지원번호')
df

df.loc['1번']   #index 1번에 해당하는 전체 데이터 

df.loc['1번','국어']   #인덱스 1번 학생의 국어성적만 가져옴

df.loc[['1번', '2번', '3번'],'영어']  #인덱스1~3번의 영어성적 

df.loc['1번':'5번', '국어':'사회']  #인덱스 1-5까지, 국어-사회까지 데이터
