==> pandas/08.DataLoc-iloc.ipynb <==
08. 데이터 선택 (iloc)
   위치를 이용하여 원하는 행에서 원하는 칼럼 선택 

import pandas as pd 
df = pd.read_csv('score.csv')
df

df.iloc[0]   #0반쩨 위치의 데이터 

df.iloc[4]

df.iloc[0:5]  #0-4번째 위치의 데이터

df.iloc[0,1]  #0번째 위치의 1번째(이름) 데이터 

df.iloc[4,2]  #5번째 학생의 1번째 데이터 

df.iloc[[0,1],[3,4]]   #0,1번째 위치의 학생의 3,4번쨰 데이터 가져오기 

df.iloc[0:5, 3:8] #0-4번째 위치의 학생 중에서, 3-7번째 데이터 가져오기
