==> pandas/09.DataCheckIf.ipynb <==
# 09.데이터 선ㅐ (조건)
조건에 해당하는 데이터 선택

import pandas as pd
df = pd.read_csv('score.csv', index_col='지원번호')
df

df['키'] >= 185   #학생들의 키가 185이상인지 여부를 true/false

filt = (df['키'] >= 185)
df[filt]

df[~filt]   #file를 역으로 적용

df[df['키'] >= 185]

df.loc[df['키'] >= 185, '수학']  #키가 185이상인 학생들의 수학 데이터 

df.loc[df['키']>=185,['이름','수학','과학']]   #키가 185인 학생의 수학, 과학 데이터 

# 다양한 조건 (and, or $, |)

df.loc[(df['키']>=185) & (df['학교'] == '북산고')]   #키가 185이상인 북산고 학생 데이터 

df.loc[(df['키'] < 170) | (df['키'] > 200)]  

# 데이터조건 (str)

filt = df['이름'].str.startswith('송')  #이름이 송으로 시작되는 데이터
df[filt]
       

filt = df['이름'].str.contains('태')  #이름에 태가 들어가는 데이터 
df[filt]

df[~filt]   #이름에 태가 들어가지 않는 데이터

langs = ["Python", "Java"]
filt = df['SW특기'].isin(langs)  #SW특기가 Python이나 Java인 데이터
df[filt]

langs = ["python","java"]
filt = df['SW특기'].str.lower().isin(langs) #SW특기가 python이나 java인 데이터
df[filt]

filt = df['SW특기'].str.contains('Java', na=True)  #NaN 데이터를 true 
df[filt]

df['SW특기'].str.contains('Java', na=False)
