==> pandas/10.EmptyData.ipynb <==
#결측치 
비어있는 데이터

import pandas as pd
df = pd.read_csv('score.csv', index_col='지원번호')
df

# 데이터 채우기 fillna

df.fillna('')   # NaN을 ''로 채움

df.fillna('없음') # NaN을 '없음'으로 채움

import numpy as np
df['학교'] = np.nan  # 학교 열을 모두 NaN으로 변경
df

df.fillna('모름')

df.fillna({'학교':'모름'})  # 학교 열의 NaN을 '모름'으로 채움

df.fillna('모름', inplace=True)  # 학교 열의 NaN을 '모름'으로 채움. 원본 데이터프레임이 변경됨
df

import pandas as pd
df = pd.read_csv('score.csv', index_col='지원번호')
df

df['SW특기'].fillna('확인 중', inplace=True) # SW특기 열의 NaN을 '확인 중'으로 채움
df

df.dropna()  # NaN이 하나라도 있는 행을 삭제
df

df.dropna(inplace=True)         # NaN이 하나라도 있는 행을 삭제. 원본 데이터프레임이 변경됨
df

df = pd.read_csv('score.csv', index_col='지원번호')
df

# axis : index or columns
# how : any or all

df.dropna(axis='index', how='any')  # NaN이 하나라도 있는 행을 삭제

df.dropna(axis='columns')   # NaN이 하나라도 있는 열을 삭제

df.dropna(axis='columns', how='all')  # 모든 값이 NaN인 열을 삭제
