==> pandas/11.Sort.ipynb <==
#데이터 정렬

import pandas as pd
df = pd.read_csv('score.csv', index_col='지원번호')
df

df.sort_values('키', ascending=False)    # 키를 기준으로 오름차순 정렬

df.sort_values(['수학','영어'], ascending=False) # 수학, 영어 점수를 기준으로 내림차순 정렬


df.sort_values(['수학','영어'], ascending=True) 

df.sort_values(['수학','영어'], ascending=[True, False])  # 수학은 오름차순, 영어는 내림차순 정렬

df['키'].sort_values(ascending=False)  # 키를 기준으로 오름차순 정렬

df.sort_index()  # index를 기준으로 오름차순 정렬

df.sort_index(ascending=False)   # 인덱스를 기준으로 내림차순 정렬
