==> pandas/12.UpdateData.ipynb <==
# 12.데이터 수정 
column 수정

import pandas as pd 
df = pd.read_csv('score.csv', index_col='지원번호')   #지원번호를 인덱스로 지정
df

df['학교'].replace({'북산고':'상북고', '능남고':'무슨고'})     #북산고를 상북고로 변경

df

df['학교'].replace({'북산고':'상북고'}, inplace=True)     #북산고를 상북고로 변경]
df

df['SW특기'].str.lower()

df['SW특기'] = df['SW특기'].str.lower()     #SW특기를 소문자로 변경
df

df['SW특기'] = df['SW특기'].str.upper()
df 

df['학교'] = df['학교'] + '등학교'     #학교에 등학교 추가 
df

#column 추가 

df['총합'] = df['국어'] + df['영어'] + df['수학'] + df['과학'] + df['사회']     #총합 추가
df

df['결과'] = 'Fail'   #결과열 추가하고 Fail로 초기화
df

df.loc[df['총합'] > 400, '결과'] = 'Pass'   #총합이 400점 이상이면 Pass로 변경
df

## column 삭제

df.drop(columns=['총합'])    #총합 열 삭제
df

df.drop(columns=['국어','수학','영어'], inplace=True)    #국어, 수학, 영어 열 삭제
df

df.drop(index='4번')    #4번 학생 데이터를 삭제 
df

filt = df['과학'] < 80  #수학 점수 80점 미만 학생 필터링
df[filt]

df[filt].index

df.drop(index=df[filt].index)

df

# Row 추가

df.loc['9번'] = ['이정환', '해남고등학교', 184, 90, 80, 'Python, C', 519, 'Pass']    #9번 학생 데이터 추가']

df

# Cell 수정

df.loc['4번','SW특기'] = 'Python, Java'    #4번 학생 SW특기 변경
df

df.loc['5번', ['학교', 'SW특기']] = ['북산고등학교', 'C, C++']    #5번 학생 학교, SW특기 변경
df 

# Column 순서 변경 

cols = list(df.columns)
cols

df = df[[cols[-1]] + cols[0:-1]]  # 결과가 맨 앞으로 오고 나머지가 그 뒤로 옴. 열 순서 변경
df

df = df[['결과', '이름', '학교']]
df

df
