==> pandas/13.Method.ipynb <==
# 13.함수적용

import pandas as pd 
df = pd.read_csv('score.csv', index_col='지원번호')
df

df['학교'] = df['학교'] + '등학교'

## 데이터에 함수 적용 (apply)

# 키에 'cm'를 붙이는 함수
def add_cm(height):
    return str(height) + 'cm' 

df['키'] = df['키'].apply(add_cm)       # 키에 'cm'를 붙이는 함수 적용
df

def capitalize(lang):
    if pd.notnull(lang):    #NaN 이 아닌지 확인
        return lang.capitalize()    #첫 글자만 대문자로 변환
    return lang

df['SW특기'] = df['SW특기'].apply(capitalize)    #SW특기에 capitalize 함수 적용
df
