==> pandas/14.grouping.ipynb <==
#14. 그룹화 
동일한 값을 가진 것들끼리 합쳐서 통계 또는 평균 등의 값을 계산하기 위해 사용

import pandas as pd 
df = pd.read_csv('score.csv', index_col = '지원번호')
df

df.groupby('학교')

df.groupby('학교').get_group('북산고')  # 북산고만 추출#

#df.groupby('학교').mean()  # 북산고의 평균값#
df.groupby('학교').size()

df.groupby('학교').size()['능남고']     #학교로 그룹화를 한 뒤 에 능남고에 해당하는 데이터의 수

df.groupby('학교').size().sort_values(ascending = False)  #학교로 그룹화를 한 뒤 각 학교별 데이터의 수를 내림차순으로 정렬

df.groupby('학교')['키'].mean()  #학교로 그룹화를 한 뒤 각 학교별 키의 평균값

df.groupby('학교')[['국어', '영어', '수학']].mean()  #학교로 그룹화를 한 뒤 각 학교별 국어, 영어, 수학의 평균값

df['학년'] = [3, 3, 2, 2, 1, 1, 3, 3]     #학년 열 추가
df

df.groupby('학교')[['이름','SW특기']].count()  #학교로 그룹화를 한 뒤 각 학교별 이름과 SW특기의 수')
df

school = df.groupby('학교')
school['학년'].value_counts()  #학교로 그룹화를 한 뒤 각 학교별 학년별 학생 수']

school['학년'].value_counts().loc['북산고'] #학교로 그룹화를 한 뒤 각 학교별 학년별 학생 수 중 북산고만 추출

school['학년'].value_counts().loc['능남고'] #학교로 그룹화를 한 뒤 각 학교별 학년별 학생 수 중 능남고만 추출

school['학교'].value_counts(normalize=True).loc['북산고']
