==> pandas/15.Quiz.ipynb <==
#Pandas 퀴즈
다음은 대한민국 영화 중에서 관객수가 가장 많은 상위 8개의 데이터입니다. 


import pandas as pd
data = {
    '영화' : ['명량', '극한직업', '신과함께-죄와 벌', '국제시장', '괴물', '도둑들', '7번방의 선물', '암살'],
    '개봉 연도' : [2014, 2019, 2017, 2014, 2006, 2012, 2013, 2015],
    '관객 수' : [1761, 1626, 1441, 1426, 1301, 1298, 1281, 1270], # (단위 : 만 명)
    '평점' : [8.88, 9.20, 8.73, 9.16, 8.62, 7.64, 8.83, 9.10]
}
df = pd.DataFrame(data)
df

1) 전체 데이터 중에서 '영화' 정보만 출력하시오

df['영화']

2) 전체 데이터 중에서 '영화', '평점' 정보를 출력하시오.

df[['영화','평점']]

3. 2015년 이후에 개봉한 영화 데이터 중에서 '영화', '개봉연도' 정보를 출력하시오.

df.loc[df['개봉 연도'] >= 2015, ['영화', '개봉 연도']]

4. 주어진 계산식을 참고하여 '추천 점수' 칼럼을 추가하시오

df['추천점수'] = (df['관객 수'] * df['평점']) // 100
df

5. 전체 데이터를 '개봉 연도' 기준 내림차순으로 출력하시오

sorted_df = df.sort_values(by='개봉 연도', ascending=False)
df
